requirements.txt
# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Geospatial processing
rasterio>=1.2.0
geopandas>=0.10.0
shapely>=1.8.0
pyproj>=3.2.0

# Earth observation
earthengine-api>=0.1.300
geemap>=0.16.0

# Visualization
matplotlib>=3.4.0
folium>=0.12.0

# AI/ML
openai>=0.27.0

# Utilities
requests>=2.26.0
python-dotenv>=0.19.0

# Optional but recommended
scikit-image>=0.18.0  # Advanced image processing
opencv-python>=4.5.0  # Computer vision
jupyter>=1.0.0        # For notebooks
tqdm>=4.62.0         # Progress bars